Move token methods from UserSession to User
