+++ src/Service/IdentityManagement/UserSession.php
@@ -4,30 +4,15 @@ namespace Lullabot\Mpx\Service\IdentityManagement;
 
 use GuzzleHttp\Promise\PromiseInterface;
 use Lullabot\Mpx\Client;
-use Lullabot\Mpx\Token;
-use Lullabot\Mpx\TokenCachePool;
 use Lullabot\Mpx\User;
 use GuzzleHttp\ClientInterface;
 use GuzzleHttp\Promise\Promise;
 use Lullabot\Mpx\Exception\ClientException;
-use Lullabot\Mpx\Exception\TokenNotFoundException;
 use Psr\Http\Message\RequestInterface;
-use Psr\Log\LoggerInterface;
-use Symfony\Component\Lock\Factory;
 use Symfony\Component\Lock\StoreInterface;
 
 class UserSession implements ClientInterface
 {
-    /**
-     * The URL to sign in a user.
-     */
-    const SIGN_IN_URL = 'https://identity.auth.theplatform.com/idm/web/Authentication/signIn';
-
-    /**
-     * The URL to sign out a given token for a user.
-     */
-    const SIGN_OUT_URL = 'https://identity.auth.theplatform.com/idm/web/Authentication/signOut';
-
     /**
      * The user to establish a session for.
      *
@@ -69,128 +54,17 @@ class UserSession implements ClientInterface
      * Note that the session is not actually established until acquireToken is
      * called.
      *
-     * @param \Lullabot\Mpx\Client         $client         The client used to access MPX.
-     * @param \Lullabot\Mpx\User           $user           The user associated with this session.
-     * @param StoreInterface               $store          The lock backend to store locks in.
-     * @param \Lullabot\Mpx\TokenCachePool $tokenCachePool The cache of authentication tokens.
-     * @param \Psr\Log\LoggerInterface     $logger         The logger used when logging automatic authentication renewals.
+     * @param \Lullabot\Mpx\Client $client The client used to access MPX.
+     * @param \Lullabot\Mpx\User   $user   The user associated with this session.
      *
      * @see \Psr\Log\NullLogger To disable logging within this session.
      */
     public function __construct(
         Client $client,
-        User $user,
-        StoreInterface $store,
-        TokenCachePool $tokenCachePool,
-        LoggerInterface $logger
+        User $user
     ) {
         $this->client = $client;
         $this->user = $user;
-        $this->tokenCachePool = $tokenCachePool;
-        $this->logger = $logger;
-        $this->store = $store;
-    }
-
-    /**
-     * Get a current authentication token for the account.
-     *
-     * This method will automatically generate a new token if one does not exist.
-     *
-     * @todo Do we want to make this async?
-     *
-     * @param int  $duration (optional) The number of seconds for which the token should be valid.
-     * @param bool $reset    Force fetching a new token, even if one exists.
-     *
-     * @return \Lullabot\Mpx\Token A valid MPX authentication token.
-     */
-    public function acquireToken(int $duration = null, bool $reset = false): Token
-    {
-        if ($reset) {
-            $this->tokenCachePool->deleteToken($this->user);
-        }
-
-        // We assume that the cache is backed by shared storage across multiple
-        // requests. In that case, it's possible for another thread to set a
-        // token between the above delete and the next try block.
-        try {
-            $token = $this->tokenCachePool->getToken($this->user);
-        } catch (TokenNotFoundException $e) {
-            $token = $this->signInWithLock($duration);
-        }
-
-        return $token;
-    }
-
-    /**
-     * Sign in the user and return the current token.
-     *
-     * @param int $duration (optional) The number of seconds for which the token should be valid.
-     *
-     * @return \Lullabot\Mpx\Token
-     */
-    protected function signIn($duration = null): Token
-    {
-        $options = [];
-        $options['auth'] = [
-            $this->user->getUsername(),
-            $this->user->getPassword(),
-        ];
-
-        // @todo Make this a class constant.
-        $options['query'] = [
-            'schema' => '1.0',
-            'form' => 'json',
-        ];
-
-        // @todo move these to POST.
-        // https://docs.theplatform.com/help/wsf-signin-method#signInmethod-JSONPOSTexample
-        if (!empty($duration)) {
-            // API expects this value in milliseconds, not seconds.
-            $options['query']['_duration'] = $duration * 1000;
-            $options['query']['_idleTimeout'] = $duration * 1000;
-        }
-
-        $response = $this->client->request(
-            'GET',
-            self::SIGN_IN_URL,
-            $options
-        );
-
-        $data = \GuzzleHttp\json_decode($response->getBody(), true);
-
-        $token = $this->tokenFromResponse($data);
-        $this->logger->info(
-            'Retrieved a new MPX token {token} for user {username} that expires on {date}.',
-            [
-                'token' => $token->getValue(),
-                'username' => $this->user->getUsername(),
-                'date' => date(DATE_ISO8601, $token->getExpiration()),
-            ]
-        );
-
-        return $token;
-    }
-
-    /**
-     * Sign out the user.
-     */
-    public function signOut()
-    {
-        // @todo Handle that the token may be expired.
-        // @todo Handle and log that MPX may error on the signout.
-        $this->client->request(
-            'GET',
-            self::SIGN_OUT_URL,
-            [
-                'query' => [
-                    'schema' => '1.0',
-                    'form' => 'json',
-                    '_token' => (string) $this->tokenCachePool->getToken($this->user),
-                ],
-            ]
-        );
-
-        $this->tokenCachePool->deleteToken($this->user);
     }
 
     /**
@@ -246,28 +120,12 @@ class UserSession implements ClientInterface
         if (!isset($options['query'])) {
             $options['query'] = [];
         }
-        $token = $this->acquireToken(null, $reset);
+        $token = $this->user->acquireToken(null, $reset);
         $options['query'] += [
             'token' => $token->getValue(),
         ];
 
         return $options;
-    }
-
-    /**
-     * Instantiate and cache a token.
-     *
-     * @param array $data The MPX signIn() response data.
-     *
-     * @return \Lullabot\Mpx\Token The new token.
-     */
-    private function tokenFromResponse(array $data): Token
-    {
-        $token = Token::fromResponseData($data);
-        // Save the token to the cache and return it.
-        $this->tokenCachePool->setToken($this->user, $token);
-
-        return $token;
     }
 
     /**
@@ -458,32 +316,5 @@ class UserSession implements ClientInterface
         });
 
         return $outer;
-    }
-
-    /**
-     * Sign in to MPX, with a lock to prevent sign-in stampedes.
-     *
-     * @param int $duration (optional) The number of seconds that the sign-in token should be valid for.
-     *
-     * @return Token
-     */
-    protected function signInWithLock(int $duration = null): Token
-    {
-        $factory = new Factory($this->store);
-        $factory->setLogger($this->logger);
-        $lock = $factory->createLock($this->user->getUsername(), 10);
-
-        // Blocking means this will throw an exception on failure.
-        $lock->acquire(true);
-
-        try {
-            // It's possible another thread has signed in for us, so check for a token first.
-            $token = $this->tokenCachePool->getToken($this->user);
-        } catch (TokenNotFoundException $e) {
-            // We have the lock, and there's no token, so sign in.
-            $token = $this->signIn($duration);
-        }
-
-        return $token;
     }
 }
