+++ tests/src/Unit/TokenCachePoolTest.php
@@ -14,7 +14,7 @@ use PHPUnit\Framework\TestCase;
 class TokenCachePoolTest extends TestCase
 {
     /**
-     * @var \Lullabot\Mpx\User
+     * @var \Lullabot\Mpx\User|\PHPUnit_Framework_MockObject_MockObject
      */
     protected $user;
 
@@ -28,7 +28,11 @@ class TokenCachePoolTest extends TestCase
      */
     public function setUp()
     {
-        $this->user = new User('username', 'password');
+        /* @var User|\PHPUnit_Framework_MockObject_MockObject $user */
+        $this->user = $this->getMockBuilder(User::class)
+            ->disableOriginalConstructor()
+            ->getMock();
+        $this->user->method('getUsername')->willReturn('username');
         $this->token = new Token('https://example.com/idm/data/User/mpx/123456', 'value', time() + 60);
     }
