+++ tests/src/Unit/UserSessionTest.php
@@ -3,10 +3,10 @@
 namespace Lullabot\Mpx\Tests\Unit;
 
 use Cache\Adapter\PHPArray\ArrayCachePool;
+use GuzzleHttp\Exception\ClientException;
 use GuzzleHttp\Exception\ServerException;
 use GuzzleHttp\Promise\PromiseInterface;
 use GuzzleHttp\Psr7\Request;
-use Lullabot\Mpx\Exception\ClientException;
 use Lullabot\Mpx\Service\IdentityManagement\UserSession;
 use Lullabot\Mpx\Tests\JsonResponse;
 use Lullabot\Mpx\Tests\MockClientTrait;
@@ -15,8 +15,6 @@ use Lullabot\Mpx\User;
 use PHPUnit\Framework\TestCase;
 use Psr\Http\Message\RequestInterface;
 use Psr\Log\LoggerInterface;
-use Psr\Log\NullLogger;
-use Symfony\Component\Lock\Exception\LockConflictedException;
 use Symfony\Component\Lock\StoreInterface;
 
 /**
@@ -26,120 +24,6 @@ use Symfony\Component\Lock\StoreInterface;
 class UserSessionTest extends TestCase
 {
     use MockClientTrait;
-
-    /**
-     * @covers ::__construct
-     * @covers ::acquireToken
-     * @covers ::signIn
-     * @covers ::signInWithLock
-     * @covers ::tokenFromResponse
-     * @covers ::signOut
-     */
-    public function testAcquireToken()
-    {
-        $client = $this->getMockClient([
-            new JsonResponse(200, [], 'signin-success.json'),
-            new JsonResponse(200, [], 'signout.json'),
-        ]);
-        $user = new User('USER-NAME', 'correct-password');
-        /** @var StoreInterface|\PHPUnit_Framework_MockObject_MockObject $store */
-        $store = $this->getMockBuilder(StoreInterface::class)
-            ->getMock();
-        $tokenCachePool = new TokenCachePool(new ArrayCachePool());
-
-        $logger = $this->fetchTokenLogger(1);
-
-        $session = new UserSession($client, $user, $store, $tokenCachePool, $logger);
-        $token = $session->acquireToken();
-        $this->assertEquals($token, $tokenCachePool->getToken($user));
-        $session->signOut();
-        $this->expectException(\RuntimeException::class);
-        $tokenCachePool->getToken($user);
-    }
-
-    /**
-     * @covers ::acquireToken
-     * @covers ::signIn
-     * @covers ::signInWithLock
-     */
-    public function testAcquireTokenFailure()
-    {
-        $client = $this->getMockClient([
-            new JsonResponse(200, [], 'signin-fail.json'),
-        ]);
-        $user = new User('USER-NAME', 'incorrect-password');
-
-        /** @var StoreInterface|\PHPUnit_Framework_MockObject_MockObject $store */
-        $store = $this->getMockBuilder(StoreInterface::class)
-            ->getMock();
-        /** @var \Psr\Log\LoggerInterface|\PHPUnit_Framework_MockObject_MockObject $logger */
-        $logger = $this->getMockBuilder(LoggerInterface::class)
-            ->getMock();
-
-        $logger->expects($this->at(0))->method('info')
-            ->with('Successfully acquired the "{resource}" lock.');
-
-        $session = new UserSession($client, $user, $store, new TokenCachePool(new ArrayCachePool()), $logger);
-        $this->expectException(ClientException::class);
-        $this->expectExceptionMessage("Error com.theplatform.authentication.api.exception.AuthenticationException: Either 'USER-NAME' does not have an account with this site, or the password was incorrect.");
-        $this->expectExceptionCode(401);
-        $session->acquireToken();
-    }
-
-    /**
-     * Test that resetting a token executes a new MPX request.
-     *
-     * @covers ::acquireToken
-     */
-    public function testAcquireReset()
-    {
-        $client = $this->getMockClient([
-            new JsonResponse(200, [], 'signin-success.json'),
-            new JsonResponse(200, [], 'signin-success.json'),
-        ]);
-        $user = new User('USER-NAME', 'correct-password');
-        /** @var StoreInterface|\PHPUnit_Framework_MockObject_MockObject $store */
-        $store = $this->getMockBuilder(StoreInterface::class)
-            ->getMock();
-        $tokenCachePool = new TokenCachePool(new ArrayCachePool());
-
-        /** @var LoggerInterface|\PHPUnit_Framework_MockObject_MockObject $logger */
-        $logger = $this->getMockBuilder(LoggerInterface::class)
-            ->getMock();
-        $logger->expects($this->at(0))->method('info')
-            ->with('Successfully acquired the "{resource}" lock.');
-        $logger->expects($this->at(1))->method('info')
-            ->with('Expiration defined for "{resource}" lock for "{ttl}" seconds.');
-        $logger->expects($this->at(2))->method('info')
-            ->willReturnCallback(function ($message, $context) {
-                $this->assertEquals('Retrieved a new MPX token {token} for user {username} that expires on {date}.', $message);
-                $this->assertArraySubset([
-                    'token' => 'TOKEN-VALUE',
-                    'username' => 'USER-NAME',
-                ], $context);
-                $this->assertRegExp('!\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}\+\d{4}!', $context['date']);
-            });
-        $logger->expects($this->at(3))->method('info')
-            ->with('Successfully acquired the "{resource}" lock.');
-        $logger->expects($this->at(4))->method('info')
-            ->with('Expiration defined for "{resource}" lock for "{ttl}" seconds.');
-        $logger->expects($this->at(5))->method('info')
-            ->willReturnCallback(function ($message, $context) {
-                $this->assertEquals('Retrieved a new MPX token {token} for user {username} that expires on {date}.', $message);
-                $this->assertArraySubset([
-                    'token' => 'TOKEN-VALUE',
-                    'username' => 'USER-NAME',
-                ], $context);
-                $this->assertRegExp('!\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}\+\d{4}!', $context['date']);
-            });
-
-        $session = new UserSession($client, $user, $store, $tokenCachePool, $logger);
-        $first_token = $session->acquireToken();
-        $this->assertEquals($first_token, $tokenCachePool->getToken($user));
-        $second_token = $session->acquireToken(null, true);
-        $this->assertEquals($second_token, $tokenCachePool->getToken($user));
-        $this->assertNotSame($first_token, $second_token);
-    }
 
     /**
      * Test normal authenticated requests.
@@ -169,7 +53,6 @@ class UserSessionTest extends TestCase
                 return new JsonResponse(200, [], 'getSelfId.json');
             },
         ]);
-        $user = new User('USER-NAME', 'correct-password');
         /** @var StoreInterface|\PHPUnit_Framework_MockObject_MockObject $store */
         $store = $this->getMockBuilder(StoreInterface::class)
             ->getMock();
@@ -177,7 +60,8 @@ class UserSessionTest extends TestCase
 
         $logger = $this->fetchTokenLogger(1);
 
-        $session = new UserSession($client, $user, $store, $tokenCachePool, $logger);
+        $user = new User($client, $store, $tokenCachePool, $logger, 'USER-NAME', 'correct-password');
+        $session = new UserSession($client, $user);
         $response = call_user_func_array([$session, $method], $args);
         if ($response instanceof PromiseInterface) {
             $response = $response->wait();
@@ -209,7 +93,6 @@ class UserSessionTest extends TestCase
             new JsonResponse(200, [], 'signin-success.json'),
             new JsonResponse(200, [], 'getSelfId.json'),
         ]);
-        $user = new User('USER-NAME', 'correct-password');
         /** @var StoreInterface|\PHPUnit_Framework_MockObject_MockObject $store */
         $store = $this->getMockBuilder(StoreInterface::class)
             ->getMock();
@@ -217,7 +100,8 @@ class UserSessionTest extends TestCase
 
         $logger = $this->fetchTokenLogger(2);
 
-        $session = new UserSession($client, $user, $store, $tokenCachePool, $logger);
+        $user = new User($client, $store, $tokenCachePool, $logger, 'USER-NAME', 'correct-password');
+        $session = new UserSession($client, $user);
         $response = call_user_func_array([$session, $method], $args);
         if ($response instanceof PromiseInterface) {
             $response = $response->wait();
@@ -246,7 +130,6 @@ class UserSessionTest extends TestCase
             new JsonResponse(200, [], 'signin-success.json'),
             new JsonResponse(403, [], '{}'),
         ]);
-        $user = new User('USER-NAME', 'correct-password');
         /** @var StoreInterface|\PHPUnit_Framework_MockObject_MockObject $store */
         $store = $this->getMockBuilder(StoreInterface::class)
             ->getMock();
@@ -254,8 +137,9 @@ class UserSessionTest extends TestCase
 
         $logger = $this->fetchTokenLogger(1);
 
-        $session = new UserSession($client, $user, $store, $tokenCachePool, $logger);
-        $this->expectException(\GuzzleHttp\Exception\ClientException::class);
+        $user = new User($client, $store, $tokenCachePool, $logger, 'USER-NAME', 'correct-password');
+        $session = new UserSession($client, $user);
+        $this->expectException(ClientException::class);
         $response = call_user_func_array([$session, $method], $args);
         if ($response instanceof PromiseInterface) {
             $response->wait();
@@ -281,7 +165,6 @@ class UserSessionTest extends TestCase
             new JsonResponse(200, [], 'signin-success.json'),
             new JsonResponse(503, [], '{}'),
         ]);
-        $user = new User('USER-NAME', 'correct-password');
         /** @var StoreInterface|\PHPUnit_Framework_MockObject_MockObject $store */
         $store = $this->getMockBuilder(StoreInterface::class)
             ->getMock();
@@ -289,38 +172,13 @@ class UserSessionTest extends TestCase
 
         $logger = $this->fetchTokenLogger(1);
 
-        $session = new UserSession($client, $user, $store, $tokenCachePool, $logger);
+        $user = new User($client, $store, $tokenCachePool, $logger, 'USER-NAME', 'correct-password');
+        $session = new UserSession($client, $user);
         $this->expectException(ServerException::class);
         $response = call_user_func_array([$session, $method], $args);
         if ($response instanceof PromiseInterface) {
             $response->wait();
         }
-    }
-
-    /**
-     * Test that acquiring a token fails if the lock cannot be grabbed.
-     *
-     * @covers ::signInWithLock
-     */
-    public function testConcurrentSignInFails()
-    {
-        $client = $this->getMockClient([
-            new JsonResponse(200, [], 'signin-success.json'),
-        ]);
-        $user = new User('USER-NAME', 'correct-password');
-        /** @var StoreInterface|\PHPUnit_Framework_MockObject_MockObject $store */
-        $store = $this->getMockBuilder(StoreInterface::class)
-            ->getMock();
-        $store->expects($this->once())->method('waitAndSave')
-            ->willThrowException(new LockConflictedException());
-        $tokenCachePool = new TokenCachePool(new ArrayCachePool());
-
-        // We cover logging in other tests.
-        $logger = new NullLogger();
-
-        $session = new UserSession($client, $user, $store, $tokenCachePool, $logger);
-        $this->expectException(LockConflictedException::class);
-        $session->acquireToken();
     }
 
     /**
